+++ PGR2project/src/SkyBox.h
@@ -25,7 +25,7 @@ public:
 
 	void init();
 
-	void update(float time);
+	void update(double time);
 
 	void translate(v3 &movVector);
